fix(test): sales invoice test cases fixes
